fix: address clippy warnings in base fee centralization
- Remove duplicate trait bounds in EthChainSpec
- Add documentation for base fee test module
- Remove unnecessary cast for INITIAL_BASE_FEE

+++ crates/ethereum/evm/src/lib.rs
@@ -230,7 +230,7 @@ where
             gas_limit *= elasticity_multiplier as u64;
 
             // set the base fee to the initial base fee from the EIP-1559 spec
-            basefee = INITIAL_BASE_FEE as u64;
+            basefee = INITIAL_BASE_FEE;
         }
 
         let block_env = BlockEnv {
